Adopt fossology_kpi_collect report to cron
Adoption is done, some minor changes have been made to the script.
First of all an instance_uuid is added to information to report.

The report is triggered by cron using pbconf/cron-db/report-kpi.sh script.
It was said that we would like to send JSON to the reporting center.
Because of this a little python script was added to report-kpi.sh file to convert from CSV to JSON.

An alternative would be to send CSV directly to Reporting API server and parse them on server side.